+++ relearn/source/Octree.cpp
@@ -768,9 +768,9 @@ void Octree::free() {
 // The caller must ensure that only inner nodes are visited. "max_level" must be chosen correctly for this
 void Octree::update_from_level(size_t max_level) {
     std::vector<double> dendrites_exc_cnts;
-    std::vector<double> dendrites_exc_connected_cnts;
+    std::vector<unsigned int> dendrites_exc_connected_cnts;
     std::vector<double> dendrites_inh_cnts;
-    std::vector<double> dendrites_inh_connected_cnts;
+    std::vector<unsigned int> dendrites_inh_connected_cnts;
 
     const FunctorUpdateNode update_node(dendrites_exc_cnts, dendrites_exc_connected_cnts, dendrites_inh_cnts, dendrites_inh_connected_cnts, 0);
